Version 3.30.4 (based on bleeding_edge revision r24443)
This uses a runtime function to set up the the constructor and its prototype (issue 3330).

Remove PersistentBase::ClearAndLeak.

Squeeze the layout of variable proxy nodes (Chromium issue 417697).

Add MonotonicallyIncreasingTime to V8 Platform (Chromium issue 417668).

Fix representation of HLoadRoot (Chromium issue 419036).

Performance and stability improvements on all platforms.

git-svn-id: https://v8.googlecode.com/svn/trunk@24444 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/compiler/x64/code-generator-x64.cc
@@ -651,7 +651,7 @@ void CodeGenerator::AssembleArchBoolean(Instruction* instr,
   // Materialize a full 64-bit 1 or 0 value. The result register is always the
   // last output of the instruction.
   Label check;
-  DCHECK_NE(0, instr->OutputCount());
+  DCHECK_NE(0, static_cast<int>(instr->OutputCount()));
   Register reg = i.OutputRegister(static_cast<int>(instr->OutputCount() - 1));
   Condition cc = no_condition;
   switch (condition) {
